data_engineering_pipeline/requirements.txt
beautifulsoup4
requests
pymongo
python-dotenv
tqdm
pytest
pandas
elasticsearch
sentence-transformers